DRAFT: adding support for enabling semantics on desktop
