+++ lib/web_ui/lib/src/engine/browser_detection.dart
@@ -121,3 +121,19 @@ OperatingSystem _detectOperatingSystem() {
     return OperatingSystem.unknown;
   }
 }
+
+/// List of Operating Systems we know to be working on laptops/desktops.
+///
+/// These devices tend to behave differently on many core issues such as events,
+/// screen readers, input devices.
+List<OperatingSystem> _desktopOperatingSystems = [
+  OperatingSystem.macOs,
+  OperatingSystem.linux,
+  OperatingSystem.windows,
+];
+
+/// A flag to check if the current operating system is a laptop/desktop
+/// operating system.
+///
+/// See [_desktopOperatingSystems].
+bool get isDesktop => _desktopOperatingSystems.contains(operatingSystem);
